requirements.txt
opencv-python==4.10.0.84
python-telegram-bot==21.3
flet==0.23.2